refactor: 리뷰 반영 PathFinder 인터페이스 적용
