+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.path.domain.Path;
+import nextstep.subway.path.domain.DijkstraPathFinder;
 import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.application.StationService;
@@ -26,7 +27,7 @@ public class PathService {
         Station sourceStation = stationService.findStationById(sourceStationId);
         Station targetStation = stationService.findStationById(targetStationId);
         List<Line> lines = lineRepository.findAll();
-        PathFinder pathFinder = new PathFinder(lines);
+        PathFinder pathFinder = new DijkstraPathFinder(lines);
         Path shortestPath = pathFinder.findShortest(sourceStation, targetStation);
         return new PathResponse(shortestPath);
     }
