+++ src/test/java/nextstep/subway/path/domain/PathFinderTest.java
@@ -49,7 +49,7 @@ class PathFinderTest {
     void findShortestTest() {
 
         //when : 양재역에서 교대역으로 가는 최단거리 검색
-        PathFinder finder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
+        PathFinder finder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
         Path path = finder.findShortest(양재역, 교대역);
 
         //then : 최단 거리와 경로를 알 수 있다.
@@ -63,7 +63,7 @@ class PathFinderTest {
     @DisplayName("같은 출발역과 도착역을 이용해 최단 거리 탐색")
     void pathFinderFailTest1() {
 
-        PathFinder finder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
+        PathFinder finder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
 
         //when : 양재역에서 양재역으로 가는 최단거리 검색
         //then : 검색 실패
@@ -78,7 +78,7 @@ class PathFinderTest {
 
         //given
         Station 도쿄역 = new Station(5L, "도쿄역");
-        PathFinder finder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
+        PathFinder finder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
 
         //when : 도쿄역에서 양재역으로 가는 최단거리 검색
         //then : 검색 실패
@@ -93,7 +93,7 @@ class PathFinderTest {
 
         //given
         Station 도쿄역 = new Station(5L, "도쿄역");
-        PathFinder finder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
+        PathFinder finder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
 
         //when : 교대역에서 도쿄역으로 가는 최단거리 검색
         //then : 검색 실패
@@ -110,7 +110,7 @@ class PathFinderTest {
         Station 도쿄역 = new Station(5L, "도쿄역");
         Station 쿄토역 = new Station(6L, "쿄토역");
         Line 신칸센 = new Line("신칸센","black", 도쿄역, 쿄토역, 40);
-        PathFinder finder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선, 신칸센));
+        PathFinder finder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선, 신칸센));
 
         //when : 교대역에서 도쿄역으로 가는 최단거리 검색
         //then : 검색 실패
